+++ frontend/src/App.tsx
@@ -5,11 +5,12 @@ import { Header } from './components/Header/Header';
 import { Footer } from './components/Footer/Footer';
 import { Messages } from './components/Messages/Messages';
 import { NotFound } from './components/NotFound/NotFound';
-import { Profile } from './components/Profile/Profile';
+import { Profile } from './components/Profile';
 import { MyMap } from './components/MyMap/MyMap';
-import { BASE_URL } from './components/Constants/Constants';
+import { BASE_URL } from './constants';
 import { IUser } from './types/Types';
 import { getCookie } from './utils/getCookie';
+import { Settings } from './components/Settings';
 
 function App() {
   const [user, setUser] = useState<IUser>();
@@ -43,8 +44,9 @@ function App() {
         <Routes>
           <Route path='/' element={<MyMap />} />
           <Route path='/messages' element={<Messages />} />
-          <Route path='/profile' element={<Profile id={userId} />} />
-          <Route path='/*' element={<NotFound />} />
+          <Route path='/settings' element={<Settings id={userId} />} />
+          <Route path='/profile/:id' element={<Profile />} />
+          <Route path='*' element={<NotFound />} />
         </Routes>
       ) : (
         <main className='main bg' />
