+++ frontend/src/components/Profile/Profile.scss
@@ -1,63 +1,6 @@
 .profile {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-grow: 1;
-}
-.form {
-  width: 500px;
-  padding: 5px;
-  @media screen and (max-width: 960px) {
-    width: 350px;
-  }
-  @media screen and (max-width: 375px) {
-    width: 280px;
-  }
-
-  .ant-form-item {
-    margin-bottom: 0;
-  }
-  .ant-form-item-label label {
-    color: white;
-  }
-  div {
-    display: block;
-  }
-  .css-dev-only-do-not-override-ixblex {
-    text-align: center;
-  }
-  .ant-select-selection-overflow {
-    display: inline;
-  }
-  button {
-    margin-top: 20px;
-    width: 200px;
-  }
-  .checkbox {
+  &__image {
     display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    justify-content: space-around;
-
-    label {
-      display: flex;
-      color: white;
-      text-align: start;
-      width: 200px;
-    }
-  }
-  .ant-input-group-addon {
-    background-color: white;
-    width: 50%;
-    text-align: end;
-  }
-  #profile_user_telegramLink {
-    text-align: start;
-  }
-}
-
-@media (max-width: 575px) {
-  .ant-form-item .ant-form-item-label {
-    padding: 0;
+    flex-direction: column;
   }
 }
